Change URLs to point to live site.  Changed local_int_t to int.
M    src/Geometry.hpp
M    README
M    INSTALL


git-svn-id: svn+ssh://software.sandia.gov/svn/hpcg/trunk@165 5e06a57e-6b39-4354-92a0-249a7d0c5410

+++ src/Geometry.hpp
@@ -22,8 +22,8 @@
 
   Define as "long long" when local problem dimension is > 2^31
 */
-//typedef int local_int_t;
-typedef long long local_int_t;
+typedef int local_int_t;
+//typedef long long local_int_t;
 
 /*!
   This defines the type for integers that have global dimension
